+++ robot-server/robot_server/service/routers/session.py
@@ -3,9 +3,9 @@ from uuid import uuid4
 
 from starlette import status as http_status_codes
 from fastapi import APIRouter, Query, Depends
-from opentrons.calibration.check.session import \
-    (CheckCalibrationSession, SessionManager,
-     CalibrationSession, CalibrationException)
+from opentrons.calibration.check.session import CheckCalibrationSession
+from opentrons.calibration.session import CalibrationSession, \
+    CalibrationException, SessionManager
 from opentrons.calibration.util import StateMachineError
 from opentrons.calibration.check import models
