+++ robot-server/tests/service/session/test_session_status.py
@@ -2,8 +2,8 @@ from unittest.mock import MagicMock
 from uuid import uuid4
 
 from opentrons.types import Mount
-from opentrons.calibration.check.session import \
-    CheckCalibrationSession, PipetteStatus, LabwareInfo
+from opentrons.calibration.check.session import CheckCalibrationSession
+from opentrons.calibration.helper_classes import PipetteStatus, LabwareInfo
 
 from robot_server.service.session.session_status import\
     _create_calibration_check_session_details
